requirements.txt
psycopg2-binary
streamlit
plotly-express